상품id, back이름 --> server로
