--- back/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "back",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon --exec babel-node app.js --delay 2"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@babel/core": "^7.14.0",
-    "@babel/node": "^7.13.13",
-    "@babel/preset-env": "^7.14.0",
-    "cookie-parser": "^1.4.5",
-    "dotenv": "^9.0.2",
-    "ejs": "^3.1.6",
-    "express": "^4.17.1",
-    "express-session": "^1.17.1",
-    "multer": "^1.4.2",
-    "mysql": "^2.18.1",
-    "passport": "^0.4.1",
-    "passport-kakao": "^1.0.1",
-    "passport-naver": "^1.0.6"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.7"
-  }
-}
