--- back/routes.js
@@ -1,23 +1,0 @@
-//Global
-const HOME = "/";
-const JOIN = "/join";
-const LOGIN = "/login";
-const LOGOUT = "/logout";
-//Users
-const USERS = "/users";
-
-//Products
-const PRODUCTS = "/products";
-const REGISTER = "/register";
-
-const routes = {
-    home : HOME,
-    join : JOIN,
-    login : LOGIN,
-    logout : LOGOUT,
-    users : USERS,
-    products : PRODUCTS,
-    register : REGISTER
-}
-
-export default routes;
