+++ client/html/insertproduct.html
@@ -4,18 +4,21 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
+    <link rel="stylesheet" href="../style/navbar.css" />
   </head>
   <body>
-    <div class="insert__product">
-      <input type="text" class="title" /><br />
-      <input type="file" class="image" /><br />
-      <input type="text" class="detail" /><br />
-      <input type="number" class="number" />
-      <input type="submit" value="상품등록" class="button" />
-    </div>
-    <a href="/product">
-      <input type="button" value="상품페이지로 가기" />
-    </a>
+    <main>
+      <div class="insert__product">
+        <input type="text" class="title" /><br />
+        <input type="file" class="image" /><br />
+        <input type="text" class="detail" /><br />
+        <input type="number" class="number" />
+        <input type="submit" value="상품등록" class="button" />
+      </div>
+      <a href="product.html">
+        <input type="button" value="상품페이지로 가기" />
+      </a>
+    </main>
     <script type="module" src="../src/insertproduct.js"></script>
   </body>
 </html>
