+++ client/html/login.html
@@ -8,9 +8,6 @@
     <link rel="stylesheet" href="../style/login.css" />
   </head>
   <body>
-    <header class="header">
-      <div class="navbar"></div>
-    </header>
     <main>
       <a href="/login/naver" class="naver">
         <img class="naver__login" src="../assets/naver.PNG" /> </a
@@ -19,5 +16,6 @@
         <img class="kakao__login" src="../assets/kakao.png" />
       </a>
     </main>
+    <script type="module" src="../src/login.js"></script>
   </body>
 </html>
