+++ client/html/main.html
@@ -17,7 +17,8 @@
               <span id="color" class="mode">Color</span></a>
             </div>
             <div class="right">
-              <a href="/login"><span>Sign in</span></a>
+              <a id="left"></a>
+              <a href="/login" id="right">Sign in</a>
             </div>
           </div>
           <div class="middle__banner">
@@ -48,6 +49,7 @@
       </header>
       <main class="document">
         <h1>상품입니다</h1>
+        <div class="products"></div>
         <div id="items" class="items">
           <a href="/product"><span>>>상품 페이지로 가기...</span></a>
         </div>
@@ -60,5 +62,5 @@
       </footer>
     </div>
   </body>
-  <script src="../src/mainpage.js"></script>
+  <script type="module" src="../src/mainpage.js"></script>
 </html>
