+++ client/html/mypage.html
@@ -14,12 +14,12 @@
         <h1 class="circle_title">TOGETOR</h1>
         <h5 class="middle_title">함께 구매해요</h5>
       </div>
+      <main class="main">
+        <div class="products"></div>
+      </main>
+      <footer class="footer">푸터입니다</footer>
     </div>
-    <main class="main">
-      <div class="products"></div>
-    </main>
-    <footer class="footer">푸터입니다</footer>
-    <script type="module" src="../src/product.js"></script>
+    <div class="no_user hidden">No User...</div>
     <script type="module" src="../src/mypage.js"></script>
   </body>
 </html>
