+++ client/html/product.html
@@ -16,7 +16,7 @@
           <li id="3" class="category">카테고리3</li>
         </ul>
       </div>
-      <a href="./insertproduct">
+      <a href="/insertproduct">
         <input type="button" value="상품등록" />
       </a>
       <div class="js__products">
