+++ client/style/mypage.css
@@ -7,22 +7,39 @@ body::-webkit-scrollbar {
 }
 
 body {
-  max-width: 1980px;
   display: flex;
   flex-direction: column;
   margin: 0;
   background-color: black;
 }
 
+.hidden {
+  display: none;
+}
+
+.no_user {
+  position: fixed;
+  top: 460px;
+  left: 810px;
+  color: white;
+  font-size: 30px;
+}
+
 .app {
   height: 1009px;
+}
+
+.js-app {
+  position: relative;
+  width: 100vw;
+  height: 100vh;
 }
 
 .circle {
   position: fixed;
   display: flex;
   top: 200px;
-  left: 650px;
+  left: 580px;
   width: 600px;
   height: 600px;
   background: var(--color);
@@ -35,7 +52,7 @@ body {
 .circle_title {
   position: fixed;
   top: 290px;
-  left: 600px;
+  left: 530px;
   color: white;
   font-size: 150px;
 }
@@ -44,7 +61,7 @@ body {
   width: 160px;
   position: fixed;
   top: 550px;
-  left: 880px;
+  left: 810px;
   color: white;
   font-size: 25px;
   font-weight: 500;
@@ -57,30 +74,28 @@ body {
 
 .main {
   position: relative;
-  width: 100vw;
   height: 100vh;
   display: flex;
   justify-content: center;
   align-items: center;
   align-content: center;
-  background-color: black;
+  background-color: rgb(255, 255, 255);
 }
 
 .footer {
   position: relative;
-  width: 100vw;
   height: 100vh;
   display: flex;
   justify-content: center;
   align-items: center;
   align-content: center;
-  color: white;
-  background-color: black;
+  color: rgb(0, 0, 0);
+  background-color: rgb(255, 255, 255);
 }
 
 .products {
   position: absolute;
-  color: white;
+  color: rgb(0, 0, 0);
 }
 
 .product {
@@ -89,7 +104,7 @@ body {
   justify-content: center;
   align-items: center;
   align-content: center;
-  background-color: aqua;
+  background-color: rgb(126, 123, 123);
 }
 
 @keyframes circle {
