+++ server/.env
@@ -1,0 +1,3 @@
+NAVER_ID=2wRRFAoU30s8HnihrrPX
+KAKAO_ID=e8249ed6217187e3a9da53e800a5b6aa
+KAKAO_SECRET=j66d2z6fT26DaKrbfsv9oM4MqTU1Oy4r
